Add dynamic approval workflow
This commit adds a dynamic approval workflow to the project. It includes a step to create or update the environment based on the result of the previous unit tests. If the unit tests fail, the environment is updated with a list of reviewers. If the unit tests pass, no changes are made to the environment.